Add LPC support for TMC22XX drivers
This adds support for TMC22XX drivers when using LPC based systems. It
uses the LPC Repetitive Interrupt Timer (RIT) to drive a custom software
UART implemetation. This allows the use of standard GPIO pins to talk to
UART based TMC devices. A single pin is used per device and the UART
operates in half duplex mode. To minimise system impact the UART
operates at a slow speed (9600 baud) and uses a single interrupt per
bit. When the UART is not operating the timer is disabled. The interrupt
handler takes approximately 1uS to execute on an LPC1768 (plus standard
interrupt overhead).

Potential issues
When enabled all active drivers must be TMC22XX devices.

The software UART is sensitive to timing issues. If other parts of the
system delay the RIT interrupt handler then errors will occur. To
improve detection of these errors full CRC checking of responses has
been enabled. Currently no errors are seen when the system is idle, but
during test print jobs an error rate of 0.5% is seen. If there is an
error the operation will be retried. Currently RepRap
firmware polls the drivers constantly to obtain status information. So
far when testing no error has created a false status report. But this is
possible if errors are not detected by the CRC code. It may make sense to change
things so that this polling is optional (can probably be done using the
current M569 R option), the UART interface would then in effect only be
used for configuration.