Switch to using file position for move system sync
This change fixes a problem seen when performing pause/resume operations
when RRF is compiled with SUPPORT_ASYNC_MOVES enabled. The existing code
made use of line numbers to track the current position within the gcode
file used by the two motion systems. However this has a number of
problems. Primarily when performing pause/resume the correct line number
is not restored, the file position is set back to the start of the last
gcode operation after the pause, but it is not easy/possible to correct
the line number beacsue of issues with lines that contain multiple gcode
statements and/or which have explicit line numbers. This problem results
in SyncWith executing the "Other input channel has skipped this sync
point" case which it should never do. This results in unpredictable
actions including spurious moves.

The solution here is to switch the sync mechanism to use the file
position rather than the line number.